requirements-testing.txt
django-stubs==1.12.0
mypy==0.971
hypothesis>=6.131.9,<7
hypothesis-jsonschema==0.19.0
types-requests==2.28.5
